requirements.txt
python-dotenv==1.0.1
Flask==3.0.3
openai==1.53.0
